refactor(docs): add `Docs` prefix to Example and Code
