+++ packages/docs/pages/components/va-modal/Modal.vue
@@ -8,19 +8,19 @@
       <h3>Basic element</h3>
       <p>This is the basic modal component implementation</p>
     </va-content>
-    <Example value="va-modal/Basic" />
+    <DocsExample value="va-modal/Basic" />
   </div>
 </template>
 
 <script>
 import Vue from 'vue'
 import VaContent from '../../../../ui/src/components/vuestic-components/va-content/VaContent'
-import Example from '../../../components/Example'
+import DocsExample from '../../../components/DocsExample'
 
 export default Vue.extend({
   components: {
     VaContent,
-    Example,
+    DocsExample,
   },
 })
 </script>
